% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicdep_data.R
\docType{data}
\name{hicdep_data}
\alias{hicdep_data}
\title{All HICDEP Tables (version 1.120)}
\format{
A list of 23 dataframes:
}
\source{
\url{https://hicdep.org/Wiki/v/10/pt/2}
}
\usage{
hicdep_data
}
\description{
Contains all HICDEP Tables (version 1.120) scraped from the HICDEP
\href{https://hicdep.org/Wiki/v/10/pt/2}{HICDEP homepage}.
}
\keyword{datasets}
