% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicdep_scrape.R
\name{hicdep_scrape}
\alias{hicdep_scrape}
\title{Scrape the HICDEP tables from the web}
\usage{
hicdep_scrape(url, table_name)
}
\arguments{
\item{url}{Website to scrape}

\item{table_name}{Name of the associated table}
}
\value{
list of tables scraped from the web
}
\description{
This function gets the tables from the
\href{https://hicdep.org/Wiki/v/10/pt/2}{HICDEP homepage}. HICDEP is a
is a platform that provides a standardised variable definition for HIV
cohorts used in Europe.
}
\examples{
\dontrun{
hicdep_scrape(
url = "https://hicdep.org/Wiki/v/10/pt/3/Table/86",
table_name = "tblART"
)}
}
